Changed ferry location coordinates format to float, fixed price for route-not-found case, fixed switching between two-way and one-way routes when defining same location for start and end, sorting order data via new "order" key in order form values before sending to server and when adding new location controls, fixed-lines-appearing-between-unused-locations bug in map-controls component

+++ migrations/20200326052440-create-locations-for-ferries.js
@@ -12,10 +12,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       latitude: {
-        type: Sequelize.DECIMAL(10, 6)
+        type: Sequelize.FLOAT
       },
       longitude: {
-        type: Sequelize.DECIMAL(11, 6)
+        type: Sequelize.FLOAT
       },
       created_at: {
         allowNull: false,
